requirements.txt
# ─── MealMatrix Backend – Python Requirements ─────────────────────────────────
# Install with:  pip install -r requirements.txt
# Python 3.10+  recommended

# ── Flask Core ─────────────────────────────────────────────────────────────────
Flask==3.0.3
Werkzeug==3.0.3

# ── Database ───────────────────────────────────────────────────────────────────
Flask-SQLAlchemy==3.1.1

# ── Server-Side Sessions ───────────────────────────────────────────────────────
Flask-Session==0.8.0

# ── CORS (Cross-Origin Resource Sharing) ──────────────────────────────────────
Flask-Cors==4.0.1

# ── Rate Limiting ──────────────────────────────────────────────────────────────
Flask-Limiter==3.6.0

# ── Environment Variables ──────────────────────────────────────────────────────
python-dotenv==1.0.1

# ── Phone OTP (SMS via Twilio) ─────────────────────────────────────────────────
twilio==9.2.3

# ── Email OTP (SMTP) ───────────────────────────────────────────────────────────
# smtplib is part of Python stdlib — no install needed

# ── Security / Cryptography ────────────────────────────────────────────────────
# secrets is part of Python stdlib — no install needed
